requirements.txt
Django==2.1.5
djangorestframework==3.7.7
# source /Users/kyu/.virtualenvs/pollsapi/bin/activate